refactor(zhuyin-game): full UI split — ModeSelect/ScoreBanner/PickGame/ComposeGame (Related to #1885)
- Extract zhuyinGameEngine.ts: deriveInitialAnswer, deriveFinalAnswer,
  buildPickChoices, buildComposePalette, scoreOnce, filterQuestionsForMode
- Extract ModeSelect.tsx (60L), ScoreBanner.tsx (58L),
  PickGame.tsx (167L), ComposeGame.tsx (187L)
- ZhuyinPhoneticGame.tsx slimmed 693→221 lines (orchestrator only)
- Add zhuyinGameEngine.test.ts: 21 tests covering 3 required TDD cases
  (parse_zhuyin splits, build_choices correct once, scoreOnce)
- All 43 zhuyin tests pass (22 logic + 21 engine)
- DO NOT touch zhuyinGameLogic.ts (existing)
